numerai-project/requirements.txt
pandas
numpy
scikit-learn
lightgbm
pyyaml
joblib
